Convert plugin loader file to TS
