+++ js/interface/menu.js
@@ -74,7 +74,6 @@ export class Menu {
 		}
 	}
 	hover(node, event, expand) {
-		console.trace('HOV', node, event ,expand)
 		if (node.classList.contains('focused') && !expand) {
 			if (node.classList.contains('hybrid_parent')) {
 				node.classList.remove('opened');
